make tests framework-agnostic (also, there is something wrong with PT)
